Fix the current context configuration may get lost
And again, `map` is lazy and may invoke its code object much later, in a
different dynamic context where `$*JSON-CLASS-CONFIG` is either
unavailable or is different from when and where the map Seq was created.
Since the problem occurs with positionals and associatives, which are
deserialized as a whole, using `eager` enforces the right context.